requirements.txt
ultralytics>=8.3.0
nnunetv2>=2.5
torch>=2.0
torchvision
timm
monai
segment-anything
opencv-python
pillow
scikit-learn
matplotlib
seaborn
tqdm
